requirements.txt
fastapi~=0.115.5
sqlmodel~=0.0.22
asyncpg
passlib[bcrypt]~=1.7.4
python-jose[cryptography]~=3.3.0
SQLAlchemy~=2.0.36
pydantic-settings~=2.6.1
pydantic~=2.9.2